+++ tests/nix_tests/test_osmosis_outpost.py
@@ -56,7 +56,7 @@ def test_ibc_transfer(ibc):
 
     assert receipt.status == 1
     # check gas used
-    assert receipt.gasUsed == 133680
+    assert receipt.gasUsed == 74098
 
     fee = receipt.gasUsed * evmos_gas_price